Store selected source in URL fragment, recall on page load
This modifies the JS on the tally display page to store the selected
source/camera in the URL fragment upon selection. This makes the tally
display for an individual camera directly addressable and bookmarkable.

Upon page load, it then reads the selected camera out of the URL
fragment (if present); otherwise behavior is unchanged.